[AIRFLOW-5457] Move GCS operators to core
