+++ airflow/contrib/operators/gcs_delete_operator.py
@@ -17,82 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 """
-This module contains Google Cloud Storage delete operator.
+This module is deprecated. Please use `airflow.gcp.operators.gcs`.
 """
+
 import warnings
-from typing import Optional, Iterable
 
-from airflow.contrib.hooks.gcs_hook import GoogleCloudStorageHook
-from airflow.models import BaseOperator
-from airflow.utils.decorators import apply_defaults
+# pylint: disable=unused-import
+from airflow.gcp.operators.gcs import GoogleCloudStorageDeleteOperator  # noqa
 
-
-class GoogleCloudStorageDeleteOperator(BaseOperator):
-    """
-    Deletes objects from a Google Cloud Storage bucket, either
-    from an explicit list of object names or all objects
-    matching a prefix.
-
-    :param bucket_name: The GCS bucket to delete from
-    :type bucket_name: str
-    :param objects: List of objects to delete. These should be the names
-        of objects in the bucket, not including gs://bucket/
-    :type objects: Iterable[str]
-    :param prefix: Prefix of objects to delete. All objects matching this
-        prefix in the bucket will be deleted.
-    :param gcp_conn_id: (Optional) The connection ID used to connect to Google Cloud Platform.
-    :type gcp_conn_id: str
-    :param google_cloud_storage_conn_id: (Deprecated) The connection ID used to connect to Google Cloud
-        Platform. This parameter has been deprecated. You should pass the gcp_conn_id parameter instead.
-    :type google_cloud_storage_conn_id: str
-    :param delegate_to: The account to impersonate, if any.
-        For this to work, the service account making the request must have
-        domain-wide delegation enabled.
-    :type delegate_to: str
-    """
-
-    template_fields = ('bucket_name', 'prefix', 'objects')
-
-    @apply_defaults
-    def __init__(self,
-                 bucket_name: str,
-                 objects: Optional[Iterable[str]] = None,
-                 prefix: Optional[str] = None,
-                 gcp_conn_id: str = 'google_cloud_default',
-                 google_cloud_storage_conn_id: Optional[str] = None,
-                 delegate_to: Optional[str] = None,
-                 *args, **kwargs) -> None:
-
-        if google_cloud_storage_conn_id:
-            warnings.warn(
-                "The google_cloud_storage_conn_id parameter has been deprecated. You should pass "
-                "the gcp_conn_id parameter.", DeprecationWarning, stacklevel=3)
-            gcp_conn_id = google_cloud_storage_conn_id
-
-        self.bucket_name = bucket_name
-        self.objects = objects
-        self.prefix = prefix
-        self.gcp_conn_id = gcp_conn_id
-        self.delegate_to = delegate_to
-
-        assert objects is not None or prefix is not None
-
-        super().__init__(*args, **kwargs)
-
-    def execute(self, context):
-        hook = GoogleCloudStorageHook(
-            google_cloud_storage_conn_id=self.gcp_conn_id,
-            delegate_to=self.delegate_to
-        )
-
-        if self.objects:
-            objects = self.objects
-        else:
-            objects = hook.list(bucket_name=self.bucket_name,
-                                prefix=self.prefix)
-
-        self.log.info("Deleting %s objects from %s",
-                      len(objects), self.bucket_name)
-        for object_name in objects:
-            hook.delete(bucket_name=self.bucket_name,
-                        object_name=object_name)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.gcp.operators.gcs`.",
+    DeprecationWarning,
+)
