+++ airflow/contrib/operators/gcs_download_operator.py
@@ -17,111 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 """
-This module contains Google Cloud Storage download operator.
+This module is deprecated. Please use `airflow.gcp.operators.gcs`.
 """
 
-import sys
 import warnings
-from typing import Optional
 
-from airflow.contrib.hooks.gcs_hook import GoogleCloudStorageHook
-from airflow.models import BaseOperator
-from airflow.models.xcom import MAX_XCOM_SIZE
-from airflow.utils.decorators import apply_defaults
-from airflow import AirflowException
+# pylint: disable=unused-import
+from airflow.gcp.operators.gcs import GoogleCloudStorageDownloadOperator  # noqa
 
-
-class GoogleCloudStorageDownloadOperator(BaseOperator):
-    """
-    Downloads a file from Google Cloud Storage.
-
-    If a filename is supplied, it writes the file to the specified location, alternatively one can
-    set the ``store_to_xcom_key`` parameter to True push the file content into xcom. When the file size
-    exceeds the maximum size for xcom it is recommended to write to a file.
-
-    :param bucket: The Google cloud storage bucket where the object is.
-        Must not contain 'gs://' prefix. (templated)
-    :type bucket: str
-    :param object: The name of the object to download in the Google cloud
-        storage bucket. (templated)
-    :type object: str
-    :param filename: The file path, including filename,  on the local file system (where the
-        operator is being executed) that the file should be downloaded to. (templated)
-        If no filename passed, the downloaded data will not be stored on the local file
-        system.
-    :type filename: str
-    :param store_to_xcom_key: If this param is set, the operator will push
-        the contents of the downloaded file to XCom with the key set in this
-        parameter. If not set, the downloaded data will not be pushed to XCom. (templated)
-    :type store_to_xcom_key: str
-    :param gcp_conn_id: (Optional) The connection ID used to connect to Google Cloud Platform.
-    :type gcp_conn_id: str
-    :param google_cloud_storage_conn_id: (Deprecated) The connection ID used to connect to Google Cloud
-        Platform. This parameter has been deprecated. You should pass the gcp_conn_id parameter instead.
-    :type google_cloud_storage_conn_id: str
-    :param delegate_to: The account to impersonate, if any.
-        For this to work, the service account making the request must have
-        domain-wide delegation enabled.
-    :type delegate_to: str
-    """
-    template_fields = ('bucket', 'object', 'filename', 'store_to_xcom_key',)
-    ui_color = '#f0eee4'
-
-    @apply_defaults
-    def __init__(self,
-                 bucket: str,
-                 object_name: Optional[str] = None,
-                 filename: Optional[str] = None,
-                 store_to_xcom_key: Optional[str] = None,
-                 gcp_conn_id: str = 'google_cloud_default',
-                 google_cloud_storage_conn_id: Optional[str] = None,
-                 delegate_to: Optional[str] = None,
-                 *args,
-                 **kwargs) -> None:
-        # To preserve backward compatibility
-        # TODO: Remove one day
-        if object_name is None:
-            if 'object' in kwargs:
-                object_name = kwargs['object']
-                DeprecationWarning("Use 'object_name' instead of 'object'.")
-            else:
-                TypeError("__init__() missing 1 required positional argument: 'object_name'")
-
-        if filename is not None and store_to_xcom_key is not None:
-            raise ValueError("Either filename or store_to_xcom_key can be set")
-
-        if google_cloud_storage_conn_id:
-            warnings.warn(
-                "The google_cloud_storage_conn_id parameter has been deprecated. You should pass "
-                "the gcp_conn_id parameter.", DeprecationWarning, stacklevel=3)
-            gcp_conn_id = google_cloud_storage_conn_id
-
-        super().__init__(*args, **kwargs)
-        self.bucket = bucket
-        self.object = object_name
-        self.filename = filename
-        self.store_to_xcom_key = store_to_xcom_key
-        self.gcp_conn_id = gcp_conn_id
-        self.delegate_to = delegate_to
-
-    def execute(self, context):
-        self.log.info('Executing download: %s, %s, %s', self.bucket,
-                      self.object, self.filename)
-        hook = GoogleCloudStorageHook(
-            google_cloud_storage_conn_id=self.gcp_conn_id,
-            delegate_to=self.delegate_to
-        )
-
-        if self.store_to_xcom_key:
-            file_bytes = hook.download(bucket_name=self.bucket,
-                                       object_name=self.object)
-            if sys.getsizeof(file_bytes) < MAX_XCOM_SIZE:
-                context['ti'].xcom_push(key=self.store_to_xcom_key, value=file_bytes)
-            else:
-                raise AirflowException(
-                    'The size of the downloaded file is too large to push to XCom!'
-                )
-        else:
-            hook.download(bucket_name=self.bucket,
-                          object_name=self.object,
-                          filename=self.filename)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.gcp.operators.gcs`.",
+    DeprecationWarning,
+)
