+++ airflow/contrib/operators/gcs_list_operator.py
@@ -17,92 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 """
-This module contains a Google Cloud Storage list operator.
+This module is deprecated. Please use `airflow.gcp.operators.gcs`.
 """
+
 import warnings
-from typing import Iterable, Optional
 
-from airflow.contrib.hooks.gcs_hook import GoogleCloudStorageHook
-from airflow.models import BaseOperator
-from airflow.utils.decorators import apply_defaults
+# pylint: disable=unused-import
+from airflow.gcp.operators.gcs import GoogleCloudStorageListOperator  # noqa
 
-
-class GoogleCloudStorageListOperator(BaseOperator):
-    """
-    List all objects from the bucket with the give string prefix and delimiter in name.
-
-    This operator returns a python list with the name of objects which can be used by
-     `xcom` in the downstream task.
-
-    :param bucket: The Google cloud storage bucket to find the objects. (templated)
-    :type bucket: str
-    :param prefix: Prefix string which filters objects whose name begin with
-           this prefix. (templated)
-    :type prefix: str
-    :param delimiter: The delimiter by which you want to filter the objects. (templated)
-        For e.g to lists the CSV files from in a directory in GCS you would use
-        delimiter='.csv'.
-    :type delimiter: str
-    :param gcp_conn_id: (Optional) The connection ID used to connect to Google Cloud Platform.
-    :type gcp_conn_id: str
-    :param google_cloud_storage_conn_id: (Deprecated) The connection ID used to connect to Google Cloud
-        Platform. This parameter has been deprecated. You should pass the gcp_conn_id parameter instead.
-    :type google_cloud_storage_conn_id:
-    :param delegate_to: The account to impersonate, if any.
-        For this to work, the service account making the request must have
-        domain-wide delegation enabled.
-    :type delegate_to: str
-
-    **Example**:
-        The following Operator would list all the Avro files from ``sales/sales-2017``
-        folder in ``data`` bucket. ::
-
-            GCS_Files = GoogleCloudStorageListOperator(
-                task_id='GCS_Files',
-                bucket='data',
-                prefix='sales/sales-2017/',
-                delimiter='.avro',
-                gcp_conn_id=google_cloud_conn_id
-            )
-    """
-    template_fields = ('bucket', 'prefix', 'delimiter')  # type: Iterable[str]
-
-    ui_color = '#f0eee4'
-
-    @apply_defaults
-    def __init__(self,
-                 bucket: str,
-                 prefix: Optional[str] = None,
-                 delimiter: Optional[str] = None,
-                 gcp_conn_id: str = 'google_cloud_default',
-                 google_cloud_storage_conn_id: Optional[str] = None,
-                 delegate_to: Optional[str] = None,
-                 *args,
-                 **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-
-        if google_cloud_storage_conn_id:
-            warnings.warn(
-                "The google_cloud_storage_conn_id parameter has been deprecated. You should pass "
-                "the gcp_conn_id parameter.", DeprecationWarning, stacklevel=3)
-            gcp_conn_id = google_cloud_storage_conn_id
-
-        self.bucket = bucket
-        self.prefix = prefix
-        self.delimiter = delimiter
-        self.gcp_conn_id = gcp_conn_id
-        self.delegate_to = delegate_to
-
-    def execute(self, context):
-
-        hook = GoogleCloudStorageHook(
-            google_cloud_storage_conn_id=self.gcp_conn_id,
-            delegate_to=self.delegate_to
-        )
-
-        self.log.info('Getting list of the files. Bucket: %s; Delimiter: %s; Prefix: %s',
-                      self.bucket, self.delimiter, self.prefix)
-
-        return hook.list(bucket_name=self.bucket,
-                         prefix=self.prefix,
-                         delimiter=self.delimiter)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.gcp.operators.gcs`.",
+    DeprecationWarning,
+)
