+++ airflow/gcp/operators/datastore.py
@@ -23,7 +23,7 @@ This module contains Google Datastore operators.
 from typing import Optional
 
 from airflow.gcp.hooks.datastore import DatastoreHook
-from airflow.contrib.hooks.gcs_hook import GoogleCloudStorageHook
+from airflow.gcp.hooks.gcs import GoogleCloudStorageHook
 from airflow.exceptions import AirflowException
 from airflow.models import BaseOperator
 from airflow.utils.decorators import apply_defaults
