+++ airflow/operators/gcs_to_s3.py
@@ -21,7 +21,7 @@ This module contains Google Cloud Storage to S3 operator.
 """
 import warnings
 
-from airflow.contrib.hooks.gcs_hook import GoogleCloudStorageHook
+from airflow.gcp.hooks.gcs import GoogleCloudStorageHook
 from airflow.contrib.operators.gcs_list_operator import GoogleCloudStorageListOperator
 from airflow.utils.decorators import apply_defaults
 from airflow.hooks.S3_hook import S3Hook
