+++ airflow/operators/local_to_gcs.py
@@ -21,7 +21,7 @@ This module contains operator for uploading local file to GCS.
 """
 import warnings
 
-from airflow.contrib.hooks.gcs_hook import GoogleCloudStorageHook
+from airflow.gcp.hooks.gcs import GoogleCloudStorageHook
 from airflow.models import BaseOperator
 from airflow.utils.decorators import apply_defaults
